Carmod AR/carmod_ar/server/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.0